refresh all migrations & rename accounts as house
